Add files via upload
Firstly, I am sorry, I tried my best but still cannot output 'HTML' from R markdown. Thus, I create a new R HTML file, but its suffix is 'Rhtml', not 'HTML'. Hoping it does not bother you, I will ask how to output 'HTML' next time in class.
Secondly, it would be better to make screen recordings for the exercise guidance part. In this case, I just need to check for the recording every time I did not remember how to deal with the code, instead of waiting in line to ask some simple questions. And it would be better if you can upload the tasks&codes that are needed for the next class several days earlier. It can help me to catch up with you.
Anyway, it is really an interesting course with a perfect teacher and TA. I enjoy it very much, thanks a lot~